ae: Move resetNextfullSync to be called in fewer places
Locally this was always being called immediately before attempting a full sync. By moving this to
immediately before the sync is attempted (instead of before returning the state) we can call the
reset fewer times, and its more obvious what is happening.

Also renames the channel to be a little bit more expresive, and remove the function now that it is only
called once, to remove indirection.